removed redundant folder, renamed package
