+++ src/main/java/algoritm/chapters/Chapter1.java
@@ -1,4 +1,4 @@
-package lv465.algoritm.chapter;
+package algoritm.chapters;
 
 import java.util.Objects;
 import java.util.stream.Stream;
