--- src/main/java/com/lv465/Main.java
@@ -1,4 +1,0 @@
-package com.lv465;
-
-public class Main {
-}
