--- src/main/java/com/lv465/chapters/Chapter2.java
@@ -1,4 +1,0 @@
-package com.lv465.chapters;
-
-public class Chapter2 {
-}
