--- src/main/java/com/lv465/chapters/Chapter3.java
@@ -1,4 +1,0 @@
-package com.lv465.chapters;
-
-public class Chapter3 {
-}
