--- src/main/java/lv465/algoritm/chapter/Chapter3.java
@@ -1,5 +1,0 @@
-package lv465.algoritm.chapter;
-
-public class Chapter3 {
-
-}
